requirements.txt
Flask
flask-cors
Pillow
reportlab
